@@ -13,16 +13,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~14.2.2",
-    "@angular/core": "~14.2.2",
+    "@angular/common": "~14.2.4",
+    "@angular/core": "~14.2.4",
     "@angular/fire": "^7.4.1",
-    "@angular/forms": "~14.2.2",
-    "@angular/platform-browser": "~14.2.2",
-    "@angular/platform-browser-dynamic": "~14.2.2",
-    "@angular/router": "~14.2.2",
-    "@angular/service-worker": "^14.2.2",
-    "@ionic/angular": "^6.2.7",
-    "@ionic/angular-server": "^6.2.7",
+    "@angular/forms": "~14.2.4",
+    "@angular/platform-browser": "~14.2.4",
+    "@angular/platform-browser-dynamic": "~14.2.4",
+    "@angular/router": "~14.2.4",
+    "@angular/service-worker": "^14.2.4",
+    "@ionic/angular": "^6.2.9",
+    "@ionic/angular-server": "^6.2.9",
     "@ngneat/until-destroy": "^9.2.2",
     "@sweetalert2/ngx-sweetalert2": "^11.0.0",
     "@types/twemoji-parser": "^13.1.1",
@@ -30,37 +30,39 @@
     "date-fns": "^2.29.3",
     "ng-qrcode": "^7.0.0",
     "ngx-clipboard": "^15.1.0",
+    "ngx-image-compress": "^13.1.13",
+    "ngx-mailto": "^2.0.0",
     "ngx-markdown": "^14.0.1",
     "ol": "^7.1.0",
-    "rxjs": "~7.5.6",
-    "sweetalert2": "^11.4.33",
+    "rxjs": "~7.5.7",
+    "sweetalert2": "^11.4.35",
     "tslib": "^2.4.0",
     "twemoji-parser": "^14.0.0",
     "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~14.2.3",
-    "@angular-eslint/builder": "~14.0.4",
-    "@angular-eslint/eslint-plugin": "~14.0.4",
-    "@angular-eslint/eslint-plugin-template": "~14.0.4",
-    "@angular-eslint/template-parser": "~14.0.4",
-    "@angular/cli": "~14.2.3",
-    "@angular/compiler": "~14.2.2",
-    "@angular/compiler-cli": "~14.2.2",
-    "@angular/language-service": "~14.2.2",
+    "@angular-devkit/build-angular": "~14.2.4",
+    "@angular-eslint/builder": "~14.1.2",
+    "@angular-eslint/eslint-plugin": "~14.1.2",
+    "@angular-eslint/eslint-plugin-template": "~14.1.2",
+    "@angular-eslint/template-parser": "~14.1.2",
+    "@angular/cli": "~14.2.4",
+    "@angular/compiler": "~14.2.4",
+    "@angular/compiler-cli": "~14.2.4",
+    "@angular/language-service": "~14.2.4",
     "@ionic/angular-toolkit": "^7.0.0",
     "@types/jasmine": "~4.3.0",
     "@types/jasminewd2": "~2.0.10",
-    "@types/node": "^18.7.18",
-    "@typescript-eslint/eslint-plugin": "5.37.0",
-    "@typescript-eslint/parser": "5.37.0",
-    "eslint": "^8.23.1",
+    "@types/node": "^18.7.23",
+    "@typescript-eslint/eslint-plugin": "5.38.1",
+    "@typescript-eslint/parser": "5.38.1",
+    "eslint": "^8.24.0",
     "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jsdoc": "39.3.6",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "jasmine-core": "~4.4.0",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.4.0",
+    "karma": "~6.4.1",
     "karma-chrome-launcher": "~3.1.1",
     "karma-coverage": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~3.0.3",
