@@ -7,9 +7,10 @@ import { IonicModule } from '@ionic/angular';
 import { PageLoginPageRoutingModule } from './page-login-routing.module';
 
 import { PageLoginPage } from './page-login.page';
+import { GoogleButtonComponent } from './components/google-button/google-button.component';
 
 @NgModule({
   imports: [CommonModule, FormsModule, IonicModule, PageLoginPageRoutingModule],
-  declarations: [PageLoginPage],
+  declarations: [PageLoginPage, GoogleButtonComponent],
 })
 export class PageLoginPageModule {}
