@@ -1,7 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthService } from '../shared/services/auth.service';
-import { Router } from '@angular/router';
-import { AngularFireAuth } from '@angular/fire/compat/auth';
 
 @Component({
   selector: 'app-page-login',
@@ -9,7 +6,7 @@ import { AngularFireAuth } from '@angular/fire/compat/auth';
   styleUrls: ['./page-login.page.scss'],
 })
 export class PageLoginPage implements OnInit {
-  constructor(public authService: AuthService, public auth: AngularFireAuth, public router: Router) {}
+  constructor() {}
 
   ngOnInit() {}
 }
