@@ -1,20 +1,25 @@
+import { ConfirmModalPage } from './confirm-modal/confirm-modal.page';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
-
 import { AddEventPageRoutingModule } from './add-event-routing.module';
-
 import { AddEventPage } from './add-event.page';
+import { SafeModule } from '../../shared/pipes/safe.module';
+import { MajorEventsService } from 'src/app/shared/services/major-event.service';
+import { SweetAlert2Module } from '@sweetalert2/ngx-sweetalert2';
 
 @NgModule({
   imports: [
     CommonModule,
     FormsModule,
+    ReactiveFormsModule,
     IonicModule,
-    AddEventPageRoutingModule
+    AddEventPageRoutingModule,
+    SafeModule,
+    SweetAlert2Module,
   ],
-  declarations: [AddEventPage]
+  declarations: [AddEventPage, ConfirmModalPage],
+  providers: [MajorEventsService],
 })
 export class AddEventPageModule {}
