@@ -41,9 +41,11 @@ export class EventListComponent implements OnInit, OnChanges {
           .collection<EventItem>('events', (ref) => {
             let query: any = ref;
             if (date) {
-              query = query.where('date', '>=', startOfDay(date)).where('date', '<=', endOfDay(date));
+              query = query
+                .where('eventStartDate', '>=', startOfDay(date))
+                .where('eventStartDate', '<=', endOfDay(date));
             }
-            return query.orderBy('date', 'asc');
+            return query.orderBy('eventStartDate', 'asc');
           })
           .valueChanges({ idField: 'id' })
           .pipe(trace('firestore'));
