@@ -4,21 +4,30 @@ export interface EventItem {
   name: string;
   icon: string;
   course: string;
-  date: Timestamp;
-  dateEnd: Timestamp;
-  location: {
-    lat: number;
-    lon: number;
-    description: string;
+  slotsAvailable?: number;
+  numberOfSubscriptions?: number;
+  eventStartDate: Timestamp;
+  eventEndDate: Timestamp;
+  location?: {
+    lat?: number;
+    lon?: number;
+    description?: string;
   };
-  description: string;
-  shortDescription: string;
-  youtubeCode: string;
-  id: string;
-  button: {
-    text: string;
+  description?: string;
+  shortDescription?: string;
+  youtubeCode?: string;
+  id?: string;
+  button?: {
+    text?: string;
     url: string;
   };
-  slots: number;
-  slotsAvailable: number;
+  inMajorEvent?: string;
+  eventType?: string;
+  public?: boolean;
+  issueCertificate?: boolean;
+  doublePresence?: boolean;
+  collectAttendance?: boolean;
+  eventGroup?: string[];
+  createdBy: string;
+  createdOn: Timestamp;
 }
